prediction_api/requirements-vercel.txt
# Core FastAPI and ASGI server dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Facebook Prophet and machine learning dependencies
# Note: Prophet might be too large for Vercel serverless functions
# Consider using lighter alternatives or deploy to a different platform
prophet==1.1.5
pandas==2.1.3
numpy==1.24.4

# Database and Supabase integration
supabase==2.0.3
asyncpg==0.29.0
psycopg2-binary==2.9.7

# HTTP client libraries
httpx==0.24.1
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.0

# Authentication and security
PyJWT==2.8.0
cryptography>=3.4.8

# Date and time utilities
python-dateutil==2.8.2

# Async support utilities
aiofiles>=23.0.0
anyio>=3.0.0

# System utilities - removed heavy dependencies for serverless
# psutil>=5.9.0  # Commented out as it's not needed for serverless

# Lightweight dependencies only
structlog==23.2.0